Create components for tag and headline byline
